lm_fit_uploader_cloud_function/requirements.txt
google-cloud-bigquery==3.13.0
pandas==2.0.2
numpy==1.24.3
statsmodels==0.14.0
db-dtypes==1.2.0